+++ packages/docs/src/stories/Checkbox.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Box, Text, Checkbox, CheckboxProps } from '@felipe-ignite-design-system/react'
+import { Box, Text, Checkbox, CheckboxProps } from '@felipe-ignite-ui-system/react'
 
 export default {
   title: 'Form/Checkbox',