+++ packages/docs/src/stories/Heading.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Heading, HeadingProps } from '@felipe-ignite-design-system/react'
+import { Heading, HeadingProps } from '@felipe-ignite-ui-system/react'
 
 export default {
   title: 'Typography/Heading',