+++ packages/docs/src/stories/Text.stories.tsx
@@ -1,5 +1,5 @@
 import type { Meta, StoryObj } from '@storybook/react'
-import { Text, TextProps } from '@felipe-ignite-design-system/react'
+import { Text, TextProps } from '@felipe-ignite-ui-system/react'
 
 export default {
   title: 'Typography/Text',