+++ packages/tokens/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @felipe-ignite-design-system/tokens
+# @felipe-ignite-ui-system/tokens
 
 ## 2.0.0
 
 ### Major Changes
 
-- First deploy design system from ignite-ui
+- First deploy