+++ packages/ts-config/CHANGELOG.md
@@ -1,7 +1,7 @@
-# @felipe-ignite-design-system/ts-config
+# @felipe-ignite-ui-system/ts-config
 
 ## 2.0.0
 
 ### Major Changes
 
-- First deploy design system from ignite-ui
+- First deploy